requirements.txt
fastapi
uvicorn
aiofiles
python-multipart
pydantic
#openai-whisper
#torch
#torchaudio
openai
httpx
anthropic
firebase-admin
genkit
google-genai
opencv-python